Move refactored server code to TOBEREMOVED file
