+++ controllers/argocd/hpa.go
@@ -15,9 +15,6 @@
 package argocd
 
 import (
-	"context"
-	"reflect"
-
 	autoscaling "k8s.io/api/autoscaling/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -55,56 +52,6 @@ func newHorizontalPodAutoscalerWithName(name string, cr *argoproj.ArgoCD) *autos
 
 func newHorizontalPodAutoscalerWithSuffix(suffix string, cr *argoproj.ArgoCD) *autoscaling.HorizontalPodAutoscaler {
 	return newHorizontalPodAutoscalerWithName(nameWithSuffix(suffix, cr), cr)
-}
-
-// reconcileServerHPA will ensure that the HorizontalPodAutoscaler is present for the Argo CD Server component, and reconcile any detected changes.
-func (r *ReconcileArgoCD) reconcileServerHPA(cr *argoproj.ArgoCD) error {
-
-	defaultHPA := newHorizontalPodAutoscalerWithSuffix("server", cr)
-	defaultHPA.Spec = autoscaling.HorizontalPodAutoscalerSpec{
-		MaxReplicas:                    maxReplicas,
-		MinReplicas:                    &minReplicas,
-		TargetCPUUtilizationPercentage: &tcup,
-		ScaleTargetRef: autoscaling.CrossVersionObjectReference{
-			APIVersion: "apps/v1",
-			Kind:       "Deployment",
-			Name:       nameWithSuffix("server", cr),
-		},
-	}
-
-	existingHPA := newHorizontalPodAutoscalerWithSuffix("server", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, existingHPA.Name, existingHPA) {
-		if !cr.Spec.Server.Autoscale.Enabled {
-			return r.Client.Delete(context.TODO(), existingHPA) // HorizontalPodAutoscaler found but globally disabled, delete it.
-		}
-
-		changed := false
-		// HorizontalPodAutoscaler found, reconcile if necessary changes detected
-		if cr.Spec.Server.Autoscale.HPA != nil {
-			if !reflect.DeepEqual(existingHPA.Spec, cr.Spec.Server.Autoscale.HPA) {
-				existingHPA.Spec = *cr.Spec.Server.Autoscale.HPA
-				changed = true
-			}
-		}
-
-		if changed {
-			return r.Client.Update(context.TODO(), existingHPA)
-		}
-
-		// HorizontalPodAutoscaler found, no changes detected
-		return nil
-	}
-
-	if !cr.Spec.Server.Autoscale.Enabled {
-		return nil // AutoScale not enabled, move along...
-	}
-
-	// AutoScale enabled, no existing HPA found, create
-	if cr.Spec.Server.Autoscale.HPA != nil {
-		defaultHPA.Spec = *cr.Spec.Server.Autoscale.HPA
-	}
-
-	return r.Client.Create(context.TODO(), defaultHPA)
 }
 
 // reconcileAutoscalers will ensure that all HorizontalPodAutoscalers are present for the given ArgoCD.
