+++ controllers/argocd/ingress.go
@@ -89,157 +89,6 @@ func (r *ReconcileArgoCD) reconcileIngresses(cr *argoproj.ArgoCD) error {
 	return nil
 }
 
-// reconcileArgoServerIngress will ensure that the ArgoCD Server Ingress is present.
-func (r *ReconcileArgoCD) reconcileArgoServerIngress(cr *argoproj.ArgoCD) error {
-	ingress := newIngressWithSuffix("server", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, ingress.Name, ingress) {
-		if !cr.Spec.Server.Ingress.Enabled {
-			// Ingress exists but enabled flag has been set to false, delete the Ingress
-			return r.Client.Delete(context.TODO(), ingress)
-		}
-		return nil // Ingress found and enabled, do nothing
-	}
-
-	if !cr.Spec.Server.Ingress.Enabled {
-		return nil // Ingress not enabled, move along...
-	}
-
-	// Add default annotations
-	atns := make(map[string]string)
-	atns[common.ArgoCDKeyIngressSSLRedirect] = "true"
-	atns[common.ArgoCDKeyIngressBackendProtocol] = "HTTP"
-
-	// Override default annotations if specified
-	if len(cr.Spec.Server.Ingress.Annotations) > 0 {
-		atns = cr.Spec.Server.Ingress.Annotations
-	}
-
-	ingress.ObjectMeta.Annotations = atns
-
-	ingress.Spec.IngressClassName = cr.Spec.Server.Ingress.IngressClassName
-
-	pathType := networkingv1.PathTypeImplementationSpecific
-	// Add rules
-	ingress.Spec.Rules = []networkingv1.IngressRule{
-		{
-			Host: getArgoServerHost(cr),
-			IngressRuleValue: networkingv1.IngressRuleValue{
-				HTTP: &networkingv1.HTTPIngressRuleValue{
-					Paths: []networkingv1.HTTPIngressPath{
-						{
-							Path: getPathOrDefault(cr.Spec.Server.Ingress.Path),
-							Backend: networkingv1.IngressBackend{
-								Service: &networkingv1.IngressServiceBackend{
-									Name: nameWithSuffix("server", cr),
-									Port: networkingv1.ServiceBackendPort{
-										Name: "http",
-									},
-								},
-							},
-							PathType: &pathType,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Add default TLS options
-	ingress.Spec.TLS = []networkingv1.IngressTLS{
-		{
-			Hosts: []string{
-				getArgoServerHost(cr),
-			},
-			SecretName: common.ArgoCDSecretName,
-		},
-	}
-
-	// Allow override of TLS options if specified
-	if len(cr.Spec.Server.Ingress.TLS) > 0 {
-		ingress.Spec.TLS = cr.Spec.Server.Ingress.TLS
-	}
-
-	if err := controllerutil.SetControllerReference(cr, ingress, r.Scheme); err != nil {
-		return err
-	}
-	return r.Client.Create(context.TODO(), ingress)
-}
-
-// reconcileArgoServerGRPCIngress will ensure that the ArgoCD Server GRPC Ingress is present.
-func (r *ReconcileArgoCD) reconcileArgoServerGRPCIngress(cr *argoproj.ArgoCD) error {
-	ingress := newIngressWithSuffix("grpc", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, ingress.Name, ingress) {
-		if !cr.Spec.Server.GRPC.Ingress.Enabled {
-			// Ingress exists but enabled flag has been set to false, delete the Ingress
-			return r.Client.Delete(context.TODO(), ingress)
-		}
-		return nil // Ingress found and enabled, do nothing
-	}
-
-	if !cr.Spec.Server.GRPC.Ingress.Enabled {
-		return nil // Ingress not enabled, move along...
-	}
-
-	// Add default annotations
-	atns := make(map[string]string)
-	atns[common.ArgoCDKeyIngressBackendProtocol] = "GRPC"
-
-	// Override default annotations if specified
-	if len(cr.Spec.Server.GRPC.Ingress.Annotations) > 0 {
-		atns = cr.Spec.Server.GRPC.Ingress.Annotations
-	}
-
-	ingress.ObjectMeta.Annotations = atns
-
-	ingress.Spec.IngressClassName = cr.Spec.Server.GRPC.Ingress.IngressClassName
-
-	pathType := networkingv1.PathTypeImplementationSpecific
-	// Add rules
-	ingress.Spec.Rules = []networkingv1.IngressRule{
-		{
-			Host: getArgoServerGRPCHost(cr),
-			IngressRuleValue: networkingv1.IngressRuleValue{
-				HTTP: &networkingv1.HTTPIngressRuleValue{
-					Paths: []networkingv1.HTTPIngressPath{
-						{
-							Path: getPathOrDefault(cr.Spec.Server.GRPC.Ingress.Path),
-							Backend: networkingv1.IngressBackend{
-								Service: &networkingv1.IngressServiceBackend{
-									Name: nameWithSuffix("server", cr),
-									Port: networkingv1.ServiceBackendPort{
-										Name: "https",
-									},
-								},
-							},
-							PathType: &pathType,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Add TLS options
-	ingress.Spec.TLS = []networkingv1.IngressTLS{
-		{
-			Hosts: []string{
-				getArgoServerGRPCHost(cr),
-			},
-			SecretName: common.ArgoCDSecretName,
-		},
-	}
-
-	// Allow override of TLS options if specified
-	if len(cr.Spec.Server.GRPC.Ingress.TLS) > 0 {
-		ingress.Spec.TLS = cr.Spec.Server.GRPC.Ingress.TLS
-	}
-
-	if err := controllerutil.SetControllerReference(cr, ingress, r.Scheme); err != nil {
-		return err
-	}
-	return r.Client.Create(context.TODO(), ingress)
-}
-
 // reconcileGrafanaIngress will ensure that the ArgoCD Server GRPC Ingress is present.
 func (r *ReconcileArgoCD) reconcileGrafanaIngress(cr *argoproj.ArgoCD) error {
 	ingress := newIngressWithSuffix("grafana", cr)
