+++ controllers/argocd/policyrule.go
@@ -44,83 +44,6 @@ func policyRuleForDexServer() []v1.PolicyRule {
 	}
 }
 
-func policyRuleForServer() []v1.PolicyRule {
-	return []v1.PolicyRule{
-		{
-			APIGroups: []string{
-				"*",
-			},
-			Resources: []string{
-				"*",
-			},
-			Verbs: []string{
-				"get",
-				"patch",
-				"delete",
-			},
-		},
-		{
-			APIGroups: []string{
-				"",
-			},
-			Resources: []string{
-				"secrets",
-				"configmaps",
-			},
-			Verbs: []string{
-				"create",
-				"get",
-				"list",
-				"watch",
-				"update",
-				"patch",
-				"delete",
-			},
-		}, {
-			APIGroups: []string{
-				"argoproj.io",
-			},
-			Resources: []string{
-				"applications",
-				"applicationsets",
-				"appprojects",
-			},
-			Verbs: []string{
-				"create",
-				"get",
-				"list",
-				"watch",
-				"update",
-				"delete",
-				"patch",
-			},
-		},
-		{
-			APIGroups: []string{
-				"",
-			},
-			Resources: []string{
-				"events",
-			},
-			Verbs: []string{
-				"create",
-				"list",
-			},
-		},
-		{
-			APIGroups: []string{
-				"batch",
-			},
-			Resources: []string{
-				"jobs",
-			},
-			Verbs: []string{
-				"create",
-			},
-		},
-	}
-}
-
 func policyRuleForNotificationsController() []v1.PolicyRule {
 	return []v1.PolicyRule{
 
@@ -201,58 +124,6 @@ func policyRuleForServerApplicationSourceNamespaces() []v1.PolicyRule {
 				"update",
 				"watch",
 				"delete",
-			},
-		},
-		{
-			APIGroups: []string{
-				"batch",
-			},
-			Resources: []string{
-				"jobs",
-			},
-			Verbs: []string{
-				"create",
-			},
-		},
-	}
-}
-
-func policyRuleForServerClusterRole() []v1.PolicyRule {
-	return []v1.PolicyRule{
-		{
-			APIGroups: []string{
-				"*",
-			},
-			Resources: []string{
-				"*",
-			},
-			Verbs: []string{
-				"get",
-				"delete",
-				"patch",
-			},
-		},
-		{
-			APIGroups: []string{
-				"argoproj.io",
-			},
-			Resources: []string{
-				"applications",
-			},
-			Verbs: []string{
-				"list",
-				"watch",
-			},
-		},
-		{
-			APIGroups: []string{
-				"",
-			},
-			Resources: []string{
-				"events",
-			},
-			Verbs: []string{
-				"list",
 			},
 		},
 		{
