+++ controllers/argocd/status.go
@@ -311,37 +311,6 @@ func (r *ReconcileArgoCD) reconcileStatusPhase(cr *argoproj.ArgoCD) error {
 	return nil
 }
 
-// reconcileStatusServer will ensure that the Server status is updated for the given ArgoCD.
-func (r *ReconcileArgoCD) reconcileStatusServer(cr *argoproj.ArgoCD) error {
-	status := "Unknown"
-
-	deploy := newDeploymentWithSuffix("server", "server", cr)
-	if argoutil.IsObjectFound(r.Client, cr.Namespace, deploy.Name, deploy) {
-		status = "Pending"
-
-		// TODO: Refactor these checks.
-		if deploy.Spec.Replicas != nil {
-			if deploy.Status.ReadyReplicas == *deploy.Spec.Replicas {
-				status = "Running"
-			} else if deploy.Status.Conditions != nil {
-				for _, condition := range deploy.Status.Conditions {
-					if condition.Type == appsv1.DeploymentReplicaFailure && condition.Status == corev1.ConditionTrue {
-						// Deployment has failed
-						status = "Failed"
-						break
-					}
-				}
-			}
-		}
-	}
-
-	if cr.Status.Server != status {
-		cr.Status.Server = status
-		return r.Client.Status().Update(context.TODO(), cr)
-	}
-	return nil
-}
-
 // reconcileStatusNotifications will ensure that the Notifications status is updated for the given ArgoCD.
 func (r *ReconcileArgoCD) reconcileStatusNotifications(cr *argoproj.ArgoCD) error {
 	status := "Unknown"
